Create screenshot View on test execution scope
onActivity method is basically a runnable scheduled on Main thread. This limits overriding Main looper.
If main looper is overridden in it, that causes dead lock.

This CL keeps activity reference and invoke viewProvider on screenshot test caller scope/context when inflateViewInTestScope parameter is true.
This let us overriding MainLooper for View creation when it is necessary e.g. ExpandableNotificationRow creation.
inflateViewInTestScope is false by default. This behavior is only valid when inflateViewInTestScope is set true.

Flag: NONE
Bug: 294241089
Test: presubmit and silver suite
Change-Id: I69572e7d2c3fa3125b952869bf75b9b4af2e5f84